more conversion fixes

+++ core/src/gbts_seeding/gbts_seeding_config.cpp
@@ -83,7 +83,7 @@ bool gbts_seedfinder_config::setLinkingScheme(const std::vector<std::pair<int, s
 	algo_params.max_Kappa*=ptScale;
 
 	//contianers sizes
-	nLayers = layerInfo.type.size();
+	nLayers = static_cast<unsigned int>(layerInfo.type.size());
 	
 	TRACCC_INFO("volume layer map has " << volumeToLayerMap_unordered.size() << " volumes");
 	TRACCC_INFO("The maxium volume index in the layer map is " << volumeToLayerMap.size());
